+++ public/Changelog.md
@@ -3,6 +3,7 @@
 **/give :**
 
 - suppression de l'option "Select an item" dans le select des items, si jamais cette valeur a téte modifiée
+- impossible de copier une commande s'il y a un message d'erreur du type "Material is not selected." et affichage d'une notification disant "impossible de copier une commande vide" si on essaie de copier
 
 # V2.1 (29-09-2024)